Add GPU detection functionality and refactor main script
